refactor code , remove colume block_no in checkpoint table
