+++ components/govoffchain-scheduler/src/main/java/org/cardanofoundation/ledgersync/govoffchainscheduler/service/offchain/drep/DRepRegistrationPersistServiceImpl.java
@@ -1,6 +1,5 @@
-package org.cardanofoundation.ledgersync.govoffchainscheduler.service.offchain.drepregistration;
+package org.cardanofoundation.ledgersync.govoffchainscheduler.service.offchain.drep;
 
-import java.sql.Timestamp;
 import java.util.List;
 
 import org.cardanofoundation.ledgersync.consumercommon.entity.OffChainDrepRegistration;
@@ -42,6 +41,8 @@ public class DRepRegistrationPersistServiceImpl implements OffChainProcessPersis
                 OffChainCheckpointType.DREP_REGISTRATION);
 
         long currentSlotNo = dRepRegistrationRepo.maxSlotNo().orElse(currentCheckpoint.getSlotNo());
+
+        // Limit the amount of query data per job run
         if (currentSlotNo - currentCheckpoint.getSlotNo() > MAX_TIME_QUERY) {
             currentSlotNo = currentCheckpoint.getSlotNo() + MAX_TIME_QUERY;
         }
@@ -61,7 +62,6 @@ public class DRepRegistrationPersistServiceImpl implements OffChainProcessPersis
         dRepRegistrationStoringService.insertFetchFailData(offChainFetchErrors);
 
         currentCheckpoint.setSlotNo(currentSlotNo);
-        currentCheckpoint.setUpdateTime(new Timestamp(System.currentTimeMillis()));
         offChainDataCheckpointStorage.save(currentCheckpoint);
 
         log.info("End fetching DRep Registration metadata, taken time: {} ms",
